Fix header rendering with Bacon-style badges
Header was showing garbage text ('? for he', truncated badges) due to
using win.print() with dynamic strings from bufPrint. The grapheme
field in libvaxis cells must point to static/comptime string memory.

Fixed by using writeCell character-by-character with a switch statement
that maps each byte to a static string literal. This ensures all
grapheme slices point to binary-embedded strings, not stack memory.

Header now shows: vigil | job | status | watching/paused

Also added:
- Test failure parsing infrastructure (TestFailure struct, extraction)
- Softer color palette for badges

+++ src/app.zig
@@ -286,7 +286,7 @@ pub const App = struct {
     }
 
     /// Scroll up by n lines.
-    fn scrollUp(self: *App, n: usize) void {
+    fn scrollUp(self: *App, n: u16) void {
         if (self.view.scroll >= n) {
             self.view.scroll -= n;
         } else {
@@ -296,7 +296,7 @@ pub const App = struct {
     }
 
     /// Scroll down by n lines.
-    fn scrollDown(self: *App, n: usize) void {
+    fn scrollDown(self: *App, n: u16) void {
         const max_scroll = self.report().getVisibleCount(self.view.expanded);
         self.view.scroll = @min(self.view.scroll + n, max_scroll);
         self.needs_redraw = true;
